Version 1.05
Added tab switching. If a buffer is displayed in another tab, then pressing 'g' will go to that tab and switch to the corresponding window.

Added file-encoding,file format,tab no and window no information for buffers in the list.

Bufferlist now opens at the very top with maximum width.

Fixed an issue with the 'g' navigation command